Split ChatMessageList and ChatMessage
